+++ server/resources/robotsTxt.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const { Transform, Writable } = require('stream');
 const log = require('../log.js');
+const { createTTLCache } = require('../api-util/cache.js');
 const { getRootURL } = require('../api-util/rootURL.js');
 const sdkUtils = require('../api-util/sdk.js');
 
@@ -32,30 +33,7 @@ const streamToPromise = stream => {
 
 // Time-to-live (ttl) is set to one day aka 86400 seconds
 const ttl = 86400; // seconds
-
-// This creates simple (proxied) memory cache
-const createCacheProxy = ttl => {
-  const cache = {};
-  return new Proxy(cache, {
-    // Get data for the property together with timestamp
-    get(target, property, receiver) {
-      const cachedData = target[property];
-      if (!!cachedData) {
-        // Check if the cached data has expired
-        if (Date.now() - cachedData.timestamp < ttl * 1000) {
-          return cachedData;
-        }
-      }
-      return { data: null, timestamp: cachedData?.timestamp || Date.now() };
-    },
-    // Set given value as data to property accompanied with timestamp
-    set(target, property, value, receiver) {
-      target[property] = { data: value, timestamp: Date.now() };
-    },
-  });
-};
-
-const cache = createCacheProxy(ttl);
+const cache = createTTLCache(ttl);
 
 // Fallback data if something failes with streams
 const fallbackRobotsTxt = `